move product variant to new file
